+++ src/internal/operations/test/exchange_test.go
@@ -18,7 +18,6 @@ import (
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/events"
 	evmock "github.com/alcionai/corso/src/internal/events/mock"
-	"github.com/alcionai/corso/src/internal/m365/collection/exchange"
 	"github.com/alcionai/corso/src/internal/m365/graph"
 	exchMock "github.com/alcionai/corso/src/internal/m365/service/exchange/mock"
 	exchTD "github.com/alcionai/corso/src/internal/m365/service/exchange/testdata"
@@ -27,6 +26,7 @@ import (
 	"github.com/alcionai/corso/src/internal/version"
 	"github.com/alcionai/corso/src/pkg/backup/details"
 	deeTD "github.com/alcionai/corso/src/pkg/backup/details/testdata"
+	"github.com/alcionai/corso/src/pkg/backup/metadata"
 	"github.com/alcionai/corso/src/pkg/control"
 	ctrlTD "github.com/alcionai/corso/src/pkg/control/testdata"
 	"github.com/alcionai/corso/src/pkg/count"
@@ -52,6 +52,18 @@ func TestExchangeBackupIntgSuite(t *testing.T) {
 
 func (suite *ExchangeBackupIntgSuite) SetupSuite() {
 	suite.its = newIntegrationTesterSetup(suite.T())
+}
+
+// MetadataFileNames produces the category-specific set of filenames used to
+// store graph metadata such as delta tokens and folderID->path references.
+func MetadataFileNames(cat path.CategoryType) [][]string {
+	switch cat {
+	// TODO: should this include events?
+	case path.EmailCategory, path.ContactsCategory:
+		return [][]string{{metadata.DeltaURLsFileName}, {metadata.PreviousPathFileName}}
+	default:
+		return [][]string{{metadata.PreviousPathFileName}}
+	}
 }
 
 // TestBackup_Run ensures that Integration Testing works
@@ -93,7 +105,7 @@ func (suite *ExchangeBackupIntgSuite) TestBackup_Run_exchange() {
 				return sel
 			},
 			category:      path.EventsCategory,
-			metadataFiles: exchange.MetadataFileNames(path.EventsCategory),
+			metadataFiles: MetadataFileNames(path.EventsCategory),
 		},
 	}
 	for _, test := range tests {
@@ -244,8 +256,8 @@ func testExchangeContinuousBackups(suite *ExchangeBackupIntgSuite, toggles contr
 		now        = dttm.Now()
 		service    = path.ExchangeService
 		categories = map[path.CategoryType][][]string{
-			path.EmailCategory:    exchange.MetadataFileNames(path.EmailCategory),
-			path.ContactsCategory: exchange.MetadataFileNames(path.ContactsCategory),
+			path.EmailCategory:    MetadataFileNames(path.EmailCategory),
+			path.ContactsCategory: MetadataFileNames(path.ContactsCategory),
 			// path.EventsCategory:   exchange.MetadataFileNames(path.EventsCategory),
 		}
 		container1      = fmt.Sprintf("%s%d_%s", incrementalsDestContainerPrefix, 1, now)
